project/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.5
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-countries==7.5.1
django-extensions==3.2.3
django-import-export==3.3.1
django-location-field==2.7.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupPy==1.14
numpy==1.26.0
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.1
phonenumbers==8.13.22
Pillow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
